--- src/pages/Reservations.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Reservations() {
-  return <div>Reservations</div>;
-}
-
-export default Reservations;
